update MoonBot to version 0.0.2

+++ src/STM32duino_LSM303AGR/LSM303AGR_ACC_Sensor.h
@@ -134,7 +134,7 @@ class LSM303AGR_ACC_Sensor
       dev_i2c->beginTransmission(((uint8_t)(((address) >> 1) & 0x7F)));
 
       dev_i2c->write(RegisterAddr);
-      for (int i = 0 ; i < NumByteToWrite ; i++)
+      for (uint16_t i = 0 ; i < NumByteToWrite ; i++)
         dev_i2c->write(pBuffer[i]);
 
       dev_i2c->endTransmission(true);
